ogma-core: Move definitions related to `ExprPair` type to dedicated module. Refs #395.
The module `ogma-core:Command.Common` exhibits low cohesion.
Specifically, it contain types and functions for abstracting expression
handlers (e.g., `ExprPair`), which are used outside the module. Other
functions in the same module manipulate values of different types,
making the module exhibit low cohesion.

This commit moves the types `ExprPair` and `ExprPairT`, and the function
`exprPair`, from `ogma-core:Command.Common` to a new internal module
`Data.ExprPair`, adjusting imports and uses as necessary.
